+++ pulse-binding/src/context/ext_device_restore.rs
@@ -19,7 +19,9 @@ use std::mem;
 use capi::pa_ext_device_restore_info as InfoInternal;
 use super::{ContextInternal, Context};
 use crate::{def, format};
-use crate::callbacks::{ListResult, box_closure_get_capi_ptr, callback_for_list_instance, ListInstanceCallback, MultiUseCallback};
+use crate::callbacks::{
+    ListResult, box_closure_get_capi_ptr, callback_for_list_instance, MultiUseCallback
+};
 use crate::operation::Operation;
 
 /// Stores information about one device in the device database that is maintained by
@@ -223,14 +225,6 @@ fn read_list_cb_proxy(_: *mut ContextInternal, i: *const InfoInternal, eol: i32,
     userdata: *mut c_void)
 {
     let _ = std::panic::catch_unwind(|| {
-        match callback_for_list_instance::<dyn FnMut(ListResult<&Info>)>(eol, userdata) {
-            ListInstanceCallback::Entry(callback) => {
-                assert!(!i.is_null());
-                let obj = Info::new_from_raw(i);
-                (callback)(ListResult::Item(&obj));
-            },
-            ListInstanceCallback::End(mut callback) => { (callback)(ListResult::End); },
-            ListInstanceCallback::Error(mut callback) => { (callback)(ListResult::Error); },
-        }
+        callback_for_list_instance(i, eol, userdata, Info::new_from_raw);
     });
 }