+++ pulse-binding/CHANGELOG.md
@@ -5,6 +5,7 @@
    too great a simplification to maintaining documentation to not implement immediately; the only
    negative is that links will be broken in locally generated documentation when built with an older
    Rust compiler than version 1.48.
+ * Re-implemented list-callback handling (internal change only).
 
 # 2.22.0 (December 22nd, 2020)